+++ Weather-iOS/Presentation/Presentation-Domain/Screen/Home/AreaFilter/AreaFilterViewController.swift
@@ -12,9 +12,8 @@ final class AreaFilterViewController: UIViewController {
 
     weak var delegate: AreaFilterTappedDelegate?
 
-    private let viewData = Area.allCases.map { $0 }
-
     private var viewModel: AreaFilterViewModel!
+    private var dataSource: AreaFilterDataSource!
 
     static func createInstance(viewModel: AreaFilterViewModel) -> AreaFilterViewController {
         let instance = AreaFilterViewController.instantiateInitialViewController()
@@ -28,9 +27,10 @@ final class AreaFilterViewController: UIViewController {
     }
 
     private func setupTableView() {
+        dataSource = AreaFilterDataSource(viewModel: viewModel)
         tableView.register(AreaFilterTableViewCell.xib(), forCellReuseIdentifier: AreaFilterTableViewCell.resourceName)
+        tableView.dataSource = dataSource
         tableView.delegate = self
-        tableView.dataSource = self
         tableView.rowHeight = 40
     }
 }
@@ -55,39 +55,6 @@ extension AreaFilterViewController: UITableViewDelegate {
             areaFilterCell.checkImageView.image = image
         }
 
-        delegate?.didSelectAreaFilter(areaIds: UserDefaults.areaIds)
-    }
-}
-
-extension AreaFilterViewController: UITableViewDataSource {
-
-    func tableView(
-        _ tableView: UITableView,
-        numberOfRowsInSection section: Int
-    ) -> Int {
-        viewData.count
-    }
-
-    func tableView(
-        _ tableView: UITableView,
-        cellForRowAt indexPath: IndexPath
-    ) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(
-            withIdentifier: AreaFilterTableViewCell.resourceName,
-            for: indexPath
-        )
-
-        if let item = viewData.any(at: indexPath.row),
-           let areaFilterCell = cell as? AreaFilterTableViewCell
-        {
-            let image = viewModel.areaIds.contains(indexPath.row) ?
-                Resources.Images.General.checkIn :
-                Resources.Images.General.checkOff
-
-            areaFilterCell.checkImageView.image = image
-            areaFilterCell.setup(item: item)
-        }
-
-        return cell
+        delegate?.didSelectAreaFilter(areaIds: viewModel.areaIds)
     }
 }
